Star

Motivation

Most optimization algorithms search for a single best solution, but this can be suboptimal for various reasons. First, many optimizers get stuck in a local min/maximum searching for a single best solution. Second, a lot of optimizers are unable to balance exploration and exploitation, which can lead them to overlook promising solutions or even get stuck in local optima. Third, almost all optimization algorithms are not ammenable to understanding the relationship between features and performance — they are bascially just black boxes. And finally, and arguably the biggest motivation, is that traditional optimization seeks a single best solution for a given situation. However, what works well in one situation might not be the best solution for another. Quality-diversity algorithms address these shortcomings by searching for a diverse set of high performing solutions, rather than a single solution. One of the most successful and widely used quality-diversity algorithms is Multi-dimensional Archive of Phenotypic Elites (MAP-Elites), which can be categorized as an evolutionary algorithm. MAP-Elites uses genetic operators to evolve high quality candidate solutions at each point within a discretized behavior (feature) space, which the user can choose.[1]

Package

Simplicity at its core: MAPElites.jl is designed to be simple and easy to use. Therefore, it contains implementations for a diverse but small core set of variants of MAP-Elites to cover a variety of situations without requiring hard-to-understand abstractions.

Customizable: MAPElites.jl lets you use whatever behavior descriptors or fitness functions you want. You can use any function as a fitness function as long as it returns a real number. This also makes it easy to mix and match genetic operators, behavior descriptions, and fitness functions, or write your own.

Library-agnostic: Unlike some Python implementations of quality-diversity algorithms, MAPElites.jl doesn't require you to use any specific library for arrays or tensors. You are free to use base Julia, CUDA arrays from CUDA.jl, or any other library. As long as your fitness function can evaluate your candidate solutions, you can use whatever data structure you want, be it arrays, strings, images, or anything else.

Works with Optimization.jl: The implementations in MAPElites.jl can also be called from within Optimization.jl (coming soon).

Alternatives

There are a few alternatives to MAPElites.jl for quality-diversity algorithms written in Python.

QDax: A Jax-based implementation of several quality-diversity algorithms mostly designed around reinforcement learning and robotics.

PyRibs: A barebones Python implementation that uses an ask-tell interface that also focuses on reinforcement learning.

qdpy: A generic minimal implementation of quality-diversity algroithms in Python.

CVT-MAP-Elites: A simple Python implementation of CVT-MAP-Elites.

Contents

Changelog

[Unreleased]

MAPElites.jl

Structs

Functions

Macros

Contributing

Reporting a Bug

Requesting New Features

Contributing Code

Updating or Fixing Documentation

Installation

Usage

Standard Workflow

Using with Optimizers.jl

Algorihtms

MAP-Elites

CVT-MAP-Elites

AURORA

SAIL

CMA-MAP-Elites

References

Motivation

Package

Alternatives

Contents

Index

References

Index

MAPElites.AbstractArchive

MAPElites.BehaviorDescriptorStrategy

MAPElites.FitnessStrategy

MAPElites.GridArchive

MAPElites.VariationStrategy

Base.getindex

Base.setindex!

MAPElites.find_niche

MAPElites.find_niche

MAPElites.select_best_elite

MAPElites.select_best_elite

MAPElites.select_random

MAPElites.select_random

MAPElites.select_worst_elite

MAPElites.select_worst_elite

MAPElites.setindex!

MAPElites.AbstractArchive — Type

An abstract type that all archives must be subtypes of. All subtypes must implement the findniche, Base.setindex!, selectrandom, selectbestelite, and selectworstelite methods.

MAPElites.BehaviorDescriptorStrategy — Type

An abstract type that all concrete behavior descriptor strategies must be subtypes of. In its simplest form, a behavior descriptor strategy can just return the original behavior descriptor vector it is passed. However, a behavioral descriptor strategy could also store state information or dimensionality reduction methods, for example. All subtypes must implement a describe_behavior method.

MAPElites.FitnessStrategy — Type

An abstract type that all fitness strategies must be subtypes of. Concrete fitness strategies can store state information such as iterations, hyperparemeters, or other data structures. For example, a fitness strategy could store a surrogate model to avoid evaluating a costly fitness function at every iteration. All subtypes must implement a fitness method.

MAPElites.GridArchive — Method

GridArchive(dims, lb, ub)

Initialize a grid archive.

Arguments

dims::NTuple: An NTuple of dimensions for the archive.

lb::NTuple: An NTuple of real numbers where each number is the lower bound for the correspoding axis of the behavior descriptor.

ub::NTuple: An NTuple of real numbers where each number is the upper bound for the correspoding axis of the behavior descriptor.

Examples

julia> dims, lb, ub = (3, 4), (0.0, -1.0), (1.0, 2.0)
julia> g = GridArchive(dims, lb, ub)
GridArchive{Tuple{Int64,Int64},Tuple{Float64,Float64}}((3, 4), (0.0, -1.0), (1.0, 2.0)

MAPElites.VariationStrategy — Type

An abstract type that all variation strategies must be subtypes of. Concrete variation strategies can store any state information such as iterations or hyperparameters and must implement the vary method.

Base.getindex — Method

Base.getindex(g, key...)

Get a tuple of the solution, behavior_descriptor, and fitness of the corresponding index.

Arguments

g::GridArchive: A GridArchive to retrive the data from.

key...::Tuple: A tuple corresponding to the index to retrieve data from.

Examples

julia> dims, lb, ub = (5,), (0.0,), (10.0,)
julia> g = GridArchive(dims, lb, ub)
julia> idx = 2, 1
julia> g[idx...]
(nothing, nothing, nothing)

Base.setindex! — Method

Base.setindex!(g, value, key...)

Set the solution, behavior_descriptor, and fitness of the corresponding index in a GridArchive.

Arguments

g::GridArchive: A GridArchive to retrive the data from.

value::Tuple{Any, AbstractVector{<:Real}, <Real}: The solution, behavior descriptor, and fitness function to assign to the key.

key...::Tuple: A tuple corresponding to the index to retrieve data from.

Examples

julia> dims, lb, ub = (5,), (0.0,), (10.0,)
julia> g = GridArchive(dims, lb, ub)
julia> idx = 2, 1
julia> g[idx...] = (90, [1.0, 2.0], 3.0)
julia> g[idx...]
(90, [1.0, 2.0], 3.0)

MAPElites.find_niche — Function

Find the index of the archive for the corresponding behavior descriptor.

This method must be implemented for all subtypes of AbstractArchive. When implementing a new find_niche method on a subtype of AbstractArchive, the method must return a tuple corresponding to the index in the archive.

MAPElites.find_niche — Method

find_niche(g, bd)

Get the index of the archive that corresponds to a given behavior descriptor.

Arguments

g::GridArchive: A GridArchive to get the niche from.

bd::AbstractVector{<:Real}: A real-valued vector representing a behavior descriptor.

Examples

julia> dims, lb, ub = (5,), (0.0,), (10.0,)
julia> g = GridArchive(dims, lb, ub)
julia> MAPElites.find_niche(g, [0.0])
(1,)

MAPElites.select_best_elite — Function

Get the best performing solution and its corresponding behavior descriptor and fitness from an archive.

This method may optionally be extended to new subtypes of AbstractArchive. When implementing a new selectbestelite method on a subtype of AbstractArchive, the method must return a tuple of soltuion, behavior descriptor, fitness.

MAPElites.select_best_elite — Method

select_best_elite(g)

Get a the best solution, its behavior descriptor, and fitness.

This method assumes the best solution has the highest fitness.

Arguments

g::GridArchive: A GridArchive to retrive the solution from.

Examples

julia> dims, lb, ub = (5,), (0.0,), (10.0,)
julia> g = GridArchive(dims, lb, ub)
julia> select_worst_elite(g)
[1.0, 2.0]

MAPElites.select_random — Function

Get a random solution and its corresponding behavior descriptor and fitness from an archive.

This method may optionally be extened to new subtypes of AbstractArchive. When implementing a new select_random method on a subtype of AbstractArchive, the method must return a tuple of solution, behavior descriptor, fitness.

MAPElites.select_random — Method

select_random(g)

Get a random solution, its behavior descriptor, and fitness.

Arguments

g::GridArchive: A GridArchive to retrive the solution from.

Examples

julia> dims, lb, ub = (5,), (0.0,), (10.0,)
julia> g = GridArchive(dims, lb, ub)
julia> select_random(g)
[1.0, 2.0]

MAPElites.select_worst_elite — Function

Get the worst performing solution and its corresponding behavior descriptor and fitness from an archive.

This method may optionally be extended to subtypes of AbstractArchive. When implementing a new selectworstelite method on a subtype of AbstractArchive, the method must return a tuple of soltuion, behavior descriptor, fitness.

MAPElites.select_worst_elite — Method

select_worst_elite(g)

Get a the worst solution, its behavior descriptor, and fitness.

This method assumes the worst solution has the lowest fitness.

Arguments

g::GridArchive: A GridArchive to retrive the solution from.

Examples

julia> dims, lb, ub = (5,), (0.0,), (10.0,)
julia> g = GridArchive(dims, lb, ub)
julia> select_worst_elite(g)
[1.0, 2.0]

MAPElites.setindex! — Function

Set the index of an archive.

All subtypes of AbstractArchive must implement this method.

References

1Mouret, Jean-Baptiste, and Jeff Clune. "Illuminating search spaces by mapping elites." arXiv preprint arXiv:1504.04909 (2015).